Selection engine: fix potential bug when undoing large amounts...
...of back-to-back selection operations, when a destructive whole-image
event occurs somewhere in the middle (e.g. Crop).  When this occurs, we
need to "play back" any selection events immediately following the
whole-image event, just like we do with non-destructive layer changes,
to ensure that we restore selection state accurately.  (Crop was
particularly problematic, because it makes some out-of-process changes
to selection state, by removing the active selection after it
completes.)

This was the last Undo-related bug I had in my personal bug-tracker, so
if you've been sitting on any nasty Undo-related problems, please do let
me know as I'm not planning on touching the engine again before 7.0
releases!