Fix potentially severe bug with WideCharToMultiByte API
This one's on Microsoft.  When using WideCharToMultiByte to convert
WCHAR strings to UTF-8, the API explicitly says you have two options:
you can pass -1 for the source length, and it will be treated as
null-terminated.  A side-effect of this is that a null char will always
be appended to the UTF-8 stream.  Or, you can pass a specific length,
and the resulting UTF-8 stream will *not* include a null-char.

Unfortunately, the second claim is bullshit, because sometimes
(unpredictably), WideCharToMultiByte *will* append a string null char to
an outgoing UTF-8 stream, even if you explicitly pass a valid length to
it.  This was causing unpredictable write-access violations because we
use precisely sized buffers.

The solution is simple enough - always allocate a few "just in case"
bytes on top of the size WideCharToMultiByte claims to need.  This was a
deeply frustrating bug, and apparently I'm not the first to encounter it
(as noted in the code).  That's what I get for trusting MSDN.  :(