Fix drag/drop from Firefox
No idea when this broke, but it's probably good it did, because our new
approach is more robust.  Web browsers are a nightmare for drag/drop
because they vomit up dozens of formats, and knowing which one is "best"
is basically black magic.  PD now preferentially looks for URLs to the
source image, because then we can download the file as-is and
potentially preserve things like metadata and ICC profiles.  (I thought
we were doing this before, but Firefox found a way around it - argh.)

While I was here, I also hardened a few related code segments, since
I've seen browsers supply bad data in any number of formats.  (This way,
if we somehow can't retrieve a source URL, whatever format we *do*
manage to grab should successfully load, regardless of browser
interference.)