Automatically remove duplicate README and LICENSE files...
...after an in-place upgrade.  These can occur because I've switched to
distributing the same GitHub Readme.md and License.md files with the
official download (instead of special Readme.txt and License.txt files);
the changeover means there may be Readme.md and Readme.txt files
side-by-side after an auto-upgrade from 6.6 to 7.0.

PD now looks for this specific case, and if the old .txt files are the
precise size of the 6.6 version(s), we remove them automatically.