Potential fix for accelerator behavior in #243
(also, there's some code clean-up here as pdAccelerator was causing
obnoxious variable casing quirks due to shared names between functions
and type members - argh)

Thank you to @jpbro for bringing this to my attention and providing a
suggested fix.  PD was incorrectly ignoring accelerators that occur in
quick succession; my intent was to only block repeat accelerators of the
same type (e.g. a user holding down Ctrl+O and spamming multiple Open
File windows), but I didn't implement this correctly.  Different
accelerators were also being blocked, but they should be allowed to pass
as quickly as the user can type them.

This fix incorporates @jpbro's ideas from PR #245, but with the handling
location moved to better deal with the way the hook proc reports key
events.

Note that this is still not a perfect fix, IMO, as quickly-fired hotkeys
still risk being ignored if subsequent orders are submitted while the
program is in the midst of processing a time-consuming previous one.
For example, on a very large image, Ctrl+A immediately followed by
Ctrl+C will possibly not handle the Ctrl+C correctly, if the "Select
All" operation is still processing.

A fix for this behavior is more cumbersome, as there are situations
where we definitely want to ignore hotkeys during processing (e.g. the
user shouldn't be pressing something like Ctrl+O during a batch
operation), but at other times, it may be desirable to set up an
internal hotkey queue and once a task finishes, look for pending hotkeys
that we can still legally process.  I'll need to do some research on
potential knock-on effects from a better fix, however, so it won't be
implemented right away.