New checks for Enter/Esc keypresses on raised dialogs
This fix was brought about by #244.  (I'm not marking that issue as
*fixed* just yet, however, because I want to make sure some additional
edge-cases are handled correctly.)

PD's many user controls are designed to automatically detect ENTER and
ESC keypresses, and if the control is sited on a window with a command
bar, those keypresses will be automatically relayed to the appropriate
OK or CANCEL button.

Unfortunately, some dialogs do not automatically assign focus to a
constituent control, which means that if a dialog has been displayed but
*not yet been interacted with*, Enter/Esc keypresses simply disappear
into the void.

To remedy this, we now manually subclass raised dialogs, and perform
additional Enter/Esc detection there.