Start testing automatic updates over https
While everything appears to be working across my various test PCs, it's
entirely possible that this setup will crash and burn elsewhere
(including XP installs, depending on server security configurations).
I'm okay with that, frankly, and it's a small price to pay for safer
program updates.

Anyway, I'd like to know what I'm getting into before the next big
release, so I'm going to ship this as part of a 7.0.1 bugfix update and
see what happens.  Yay for public-facing experiments!