Modernize and streamline automatic update engine
Still some work to do here, but this is a good improvement over the old
version.  Because PD now ships updates as whole units (rather than
attempting to support piecemeal patching), we can cut down on a lot of
potentially buggy code.

Further developments will happen on the master branch, as I want these
updates to ship in 7.0.1.