Clean-up resize handling for the left toolbox
The toolbox is now forced to a multiple of the current button width
(with additional considerations for padding, obviously), which reduces
wasted space when manually resized by the user.  Similarly, changing
button width from the Window > Toolbox menu now auto-triggers a resize
to fit the new button width.

There are some weird text sizing issues at small sizes (where the
titlebar controls can't seem to pick a consistent size) but I am working
on a solution to this.  Otherwise, this approach seems to be better in
every regard.