Effects > Stylize > Palettize: support embedded palette groups
Relates to #249

To see how this works, you can download Google's Material Design palette
in Adobe's ASE format from this link:

https://material.io/guidelines/resources/color-palettes.html

Find and extract the .ASE file, then start PD, load an image, then run
the Effects > Stylize > Palettize menu.  Select the ASE file you just
extracted and voila!  All palettes embedded in the file are now
presented in a listbox.

This completes my current work on palette file parsing.  It still might
be fun to provide a dedicated "Retro" tool, one that uses a palette
collection we manually embed/ship (and potentially using a more
complicated palette matcher that adheres to original device hardware
limitations), so I'm going to leave #249 open for now, with hopes of
revisiting it in the future.