Right toolbox: new palette-based color selector
The commit title says it all: PD's right-hand "color box" now allows the
user to switch between the default color wheel and a palette-based color
selector.  The palette auto-sizes to fit the available area, which has
reminded me of the (obvious) need for dynamic vertical resizing of
panels on that right-side panel.  I'll try to add that soon.

Cool features in the current build include dynamic tooltips with color
names (extracted from all formats that support names!) when colors are
hovered.  The new control also plays nicely with existing PD color
selector features, like dynamically repainting all color selector
controls in various tool option panels when the active color changes.
The usual requirements like runtime theming, language translations, and
remembering last-used settings are obviously in place as well.

That said, there are some deficiencies that still need rectifying...

1) How to handle palettes with multiple groups (e.g. Adobe Swatch
files).  For now, I'll probably just add a listbox to the popup settings
box, but in the future, a quicker mechanism would be better.

2) The color picker tool doesn't really make sense when a palette is
active.  What I need to do is write a function to match colors selected
via the color picker against the *nearest* palette color, then
auto-select that.  This could be very useful for e.g. recoloring, as we
would dynamically restrict color picker results to the current palette.

3) Very large palettes obviously result in very small palette color
boxes.  Adding vertical resizing as an option will help with this, but
other programs tend to provide a scroll bar, as necessary.  I *really*
don't want to do this as it's hugely inconvenient for rapid color
selection (which is the whole point of using a palette in the first
place, no?)... but if enough people complain, I'll relent.