Right toolbox: add grippers to titlebars, improve reflow logic
Thank you to @Kroc for suggestions related to this recent work on
vertically resizable panels.

Titlebars now render left-side grippers when their .Draggable property
is set.  The right toolbox takes advantage of this, along with some
rewritten reflow logic, to better render the "drag process" as subpanels
are resized.  Now that subpanel windows are manually invalidated and
redrawn as the drag occurs, the entire process looks and feels very
fluid, regardless of panel size, OS version, or hardware.

No further modifications to this panel are planned, but ideas for
continued improvements are always welcome!