pdDebug: allow arbitrary start/suspend of debug logging during a session
This allows us to start or kill the debugger mid-session based on user
preference changes.  If the user chooses to activate debug logs in the
middle of a session, previously recorded debug messages will be
immediately dumped to file.  If the user chooses to deactivate debug
logs, the existing log will be immediately closed and no further log
data will be written to file.