Tooltips: improve positioning logic
Kinda relates to #253.

PD's tooltips are all positioned automatically.  The position function
attempts to position the tooltip as close as possible to the mouse
cursor, without obscuring the underlying control.

Changes in this version include:
- Potential alignments for the tooltip (left/top/right/bottom) are now
sorted from closest to furthest distance from the cursor.  If the
nearest edge of the control is not available for positioning (as it
would result in the tooltip lying off-screen), PD now moves to the
next-closest position in the list, and continues until it finds a valid
location.  (Previously, invalid positions would simply "flip" the
tooltip to the other side of the control; this resulted in poor
positioning if the control lay near the bottom-right corner of a wide
control, for example.)
- If a tooltip is already visible for the underlying window, PD now
attempts to reuse the tooltip's existing position instead of calculating
a new one.  Despite this, the tooltip position may still change to
ensure that it does not overlap the underlying control.
- For bottom-aligned tooltips, cursor size is now taken into account,
and PD positions the tooltip with enough padding that the mouse cursor
will not overlap the tooltip window, obscuring its text.
- Other minor positioning fixes and improvements.