Image > Canvas Size: remember last anchor position
That dialog is a bit weird, as it deliberately does *not* store the
last-used settings on the dialog (because we need to forcibly reset them
to match the current image's size).  However, the custom "anchor
position" control *should* store its last-used layout, as the user may
need to crop a bunch of images in the same pattern, and we don't want to
reset that between calls.

Anyway, the command bar now provides a workaround for this hybrid
requirement of "remember some last-used settings, but not others".