Drag/Drop now prompts for desired behavior
Dropped images can be opened normally, or added as a new layer to the
existing image.

(Previously, this behavior was unclear - dropping to the canvas would
create a new layer, while dropping to any non-client areas would create
a new image.  It makes much more sense to explicitly prompt for desired
behavior.)