When loading images, note if the source file contains alpha data
I'm leaning on FreeImage for this, so it may not work for all file
format/color space combinations (esoteric PNGs are particularly
problematic, as their possibility grid is about 1000x more complicated
than it needs to be).  But for the kind of files normal users have, this
should work very well.