PNG export: soft test of "use file's original palette"
Original palette can now be previewed in the PNG export screen, as a
sort of "proof of concept" that this approach could work.  Neat!

That said, to implement this feature correctly, I need to completely
overhaul PD's core export interface with FreeImage.  I've known this
would be necessary for awhile (sigh) but it's a *huge* project, and I'm
not sure I'm prepared to tackle it right now.  With the past few commits
in place, a future overhaul will be simpler to test than it would
otherwise be, but more homework is required before going down this
rabbit hole.

As such, I'm suspending work on this for now (and hiding the UI bits
that trigger this).  I'll revisit after additional investigation.