Greatly simplify cursor handling
Back when PD was using a horrifying mix of built-in VB UI elements and custom UI elements, we had to play nasty games with class cursor settings to prevent VB from forcibly changing the cursor back to whatever the design-time property dictated.  Now that we're free of that debacle, we can rely on standard WM_SETCURSOR messages.