GDI+: new interface for property retrieval, prep for multipage support
FreeImage is finicky with multipage images, and sometimes it fails for unknown reasons.  As I move more and more away from it, it'd be nice to have a GDI+ fallback for multipage TIFF files as that's currently our only real layer-friendly mechanism for interop with other software.

In this commit, the GDI+ load function has been reworked against proper property retrieval (our old interface was hackish and bad), and multipage TIFFs are now correctly identified and reported back to the calling function.  Actual implementation of multipage > layer conversion is still TODO