Image importer: overhaul TIFF loading
TIFFs are a nightmare in many different ways, and unfortunately for us, FreeImage is prone to a ton of weird issues with 'em.  I don't know if I should blame FreeImage or libTiff for this (or the TIFF spec/lack-of-spec in general), but since multipage TIFFs are currently our only real mechanism for multi-layer interop with other software, I need to improve stability.

I've now rewritten PD's GDI+ loader to support multi-page parsing, with full support for things like per-page color-management, orientation flags, and CMYK handling.  GDI+ (at least on Win 10) has proven to be far more reliable *and* fast at handling TIFF data, so it's wins all around.

Because of this, I've also rewritten PD's TIFF import strategy to first attempt loading via GDI+, and if that fails, fall back to FreeImage.  On XP, this is useful as GDI+ was underdeveloped back then.  It also gives us twice the coverage of esoteric TIFF formats, which is more than most software offers.

A lot of files were touched as part of this rewrite, but the new GDI+ loader is much cleaner (even for other file formats), so I think it's a win all-around.

As part of the work, I stripped out a bunch of superfluous translation segments, and consolidated some overwrought GDI+ interface bits.