Remove Git LFS support; rewrite history (again)
Apologies for any trouble this causes with forks, but Git LFS has proven to be an endless source of headaches.  Its integration with GitHub is remarkably poor (e.g. "Download as Zip" permanently breaks, commit titles get converted to descriptions, it is impossible to ever *remove* blobs from LFS storage), and the stringent bandwidth and space requirements are going to *cost* me money in the long run.  Long story short, if you're going to make the painful decision to forcibly rewrite your git history, there are better (and cheaper) ways to reduce repo space.

As such, I'm reverting to a repo copy prior to the Git LFS conversion, then re-applying all subsequent commits in a single "giant" commit.  I realize this is inelegant but I don't know a better way to do it.

Commits included here:
- create proper issue templates
- convert various in-app and documentation links to the new https version of photodemon.org
- new update patcher built around a dedicated GH-pages site (instead of relying on raw binary blobs inside a repo)
- new builds of some 3rd-party plugins