Undo engine: record all user behavior in macro-compatible format
Relates to #265.  Thank you to @jpbro for pointing me down this path.

In the past, PD's Undo engine only stored a subset of information that the Macro engine stores.  (For example, Undo tracks operation names, so it can display them in the Undo History browser, but it doesn't track other action attributes, like "is a dialog associated with this action?")

As of this commit, this is no longer the case.  PD's Undo engine now tracks user operations in a manner very similar to the Macro engine.  This actually results in cleaner code in many places (because we can pass around bare PD_ProcessCall structs instead of manually stripping out individual elements), and it's a prerequisite to creating macro data from arbitrary points in a given editing session.

While I was here, I also caught and fixed a few minor bugs...
- The "Modify selection" Undo menu text - used when the user makes a non-destructive setting change via something other than the canvas, e.g. toggling a selection tool dropdown - was not being captured by the translation engine.
- Rectangular selections *without* feathering were sometimes creating two Undo stack entries instead of one, because selection dimensions were being cast to integer in some places but not others.  (So one action was being created for the floating-point coords, and a second one for the integer coords - ugh!)

This overhaul required some fairly deep changes to the Undo/Redo engine tracks data, so please let me know if you run into any unexpected behavior.