Implement Tools > Create macro > From session history menu
Relates to #265.

This is my first take on @jpbro's suggestion for an "apply history to some other image" feature.  I've placed it in a new Tools > Create Macro menu (alongside macro recording) as that seemed most intuitive to me, but I am open to feedback/other suggestions!

The code was quick to piece together as it borrows liberally from the Undo History window.  Two listboxes are provided: one for the user to select a "first action" from the current session's history, and another to select the "last action".  All actions between "first" and "last" (inclusive) are then auto-converted to PD's macro format and handed off to the macro engine.

In a rare bit of proper code reuse, the existing Macro engine has been reworked to accept external action lists, which it then writes using the same exporter as recorded macros.  I've used this to verify that whether recording a macro, or exporting it retroactively via this new tool, the resulting macro file is byte-for-byte identical.

I wouldn't be surprised if many people actually prefer this tool to the old method of recording macros, as this plays much better with the typical "experimenting as you go" approach to photo editing.

Many thanks to @jpbro for this suggestion.  I'm going to leave #265 open for a bit longer (pending feedback and some additional testing) but hopefully the bulk of the work is ready for end users.