Evaluate basic expressions in all edit boxes
Relates to #263.  Many thanks to @jpbro and @vbRichClient for making this possible.

PhotoDemon will now attempt to evaluate mathematical expressions entered into any pdSpinner control.  (That's the control that looks like a text box, updown, and "reset" button crammed into one.  It's also a component of every slider control.)

+ - * / ^ and % (mod) are currently supported.  Other operators and/or functions can be added if people need 'em.  Parentheses and order of operations are respected.  Parentheses that are not paired correctly (e.g. an opening parenthesis without a closing one) are resolved automatically, by acting as if the missing symbol exists at the start or end of the expression (whichever is required).  This means that "2 * (3 + 4" evaluates identically to "2 * (3 + 4)", which is important so that PD can return evaluation results *as* the user is typing.

Spinner controls will automatically update their .Value property as an expression is entered, but they will *not* update the edit box contents with the final result until the edit box loses focus.  In adjustment/effect windows, this allows you to see what the adjustment/effect looks like "as you type", without requiring you to switch away from the edit box to force an evaluation.

Finally, the spinner's existing UI for invalid entries is still in place - specifically, a chunky red border is displayed if the current formula is un-evaluateable.  Many thanks to @jpbro for the "CanEvaluate" function that makes this possible!  (Similarly, expressions that evaluate to a value outside the range of the control are still caught and reported correctly, I believe.)

Notes to self:
- a nice additional enhancement would be a tooltip that automatically pops up and displays the expression followed by "=<result>", if the control detects that the user is entering a formula