Formulas in edit boxes: insta-display tooltip with expression result
Since spin controls tend to be quite small horizontally, it's easy to overflow them when entering a formula or expression.  Non-numeric entries now auto-trigger a tooltip display, which shows both the entire entered formula, and the calculated result (if the formula is valid).

Invalid entries (whether numeric or full formulas) also get an auto-displayed tooltip, one that is custom-tailored to the particular error.  For example, out-of-range problems display differently than "invalid formula" ones.  This should make it trivial for users to correct any problems as-they-go.