Continued work on cairo wrapper
Scaling via cairo has proven to be a bust, with performance (at similar quality settings) coming in 2x slower than GDI+.  I may not be using cairo optimally, and its API is unpleasantly complex for simple tasks like "emulate StretchBlt", but that's okay - image resizing isn't meant to be its primary use-case.  Mostly I was just hoping for an easy performance win, but alas, GDI+ has proven to be the continuing winner in that regard.