Fix competing window order between dropdown and tooltips
This was most apparent when raising the Typography tool's Font Selection dropdown (which is quite tall), and accidentally mousing over the color picker controls - because they now raise tooltips immediately, even when they don't have focus, they were competing with the dropdown for top-level window status.

Several solutions are possible; I've opted for simply not showing tooltips on underlying controls when a dropdown is raised.