Implement automatic spritesheets for caching some UI elements
At present, only the pdButtonToolbox control (used for toolbox buttons, primarily) is using this new UIImages cache.  I want to roll this change out slowly, as it potentially affects a ton of UI code, and it would be easy to accidentally break things.

The new UIImages cache is basically an automatic spritesheet generator.  When you use it to load an image resource, it automatically adds the resource to a spritesheet, and instead of returning a bare DIB, it returns an index into the spritesheet.  You can then pass the index to a UIImages.PaintCachedImage() function which handles the messy details of converting indices into actual image coordinates, and painting them accordingly.

On a cold startup, this cuts PD's GDI object count by a solid 10%, which is great!  Rolling this feature out to additional places will allow me to continue to reduce object usage, as necessary.  (The biggest gain will come from caching menu icons, but unfortunately, that requires us to go fully owner-drawn as the "built-in" menu icons feature in Windows require you to supply a unique DIB for each menu icon... ugh.)

I believe I've properly covered all the weird circumstances that may lead to broken UI images, including live-changing the UI theme, but if you run into any oddities, please let me know.

(Also included in this commit is an updated German language file from Roy K - thank you, Roy!)