Fix potential crash when the Choose Theme dialog is canceled
The new UIImages cache will have already cleared its internal cache (it must in order to live-update the "choose theme" dialog properly), but it was only rebuilding its cache if OK was selected.  This subsequently breaks everything if CANCEL was selected, but a fix is simple - the cache needs to be forcibly rebuilt *regardless* of how the dialog is closed.