FreeImage: update to 3.18 official
After 3.5 years, FreeImage has [finally released an official update](http://freeimage.sourceforge.net/news.html).

I've been custom-building PD's copy in the interim as it needed a host of important bug- and vuln-fixes, but now that an official release is out, I'm happy to ship the official version instead.  (This also seems to help with erroneous virus scanner results, as FreeImage is a common target for false-positives, and their "official" build gets more people self-reporting it as safe.)