JPEGs: improve load speed
On old versions of Windows, it was preferable to use FreeImage for JPEG import as the GDI+ loader was prone to problems on esoteric JPEG types (e.g. CMYK).  GDI+ is improved as of Win 8+ however and is faster than FreeImage due to FreeImage's obnoxious insistence on packaging libjpeg.

This simple change cuts the JPEG import time by about 50% by my testing.