Gradient tool: first draft
Tool UI is just a placeholder for now and of course many options are not yet exposed (e.g. gradient shape).  However, the current build is fully functional and expected features like clipping to the active selection are already in place.

That said, performance for additional features will be a problem, frankly a bigger problem than I anticipated.  Linear gradients are nicely accelerated (as seen in this build) and I have even managed to work around GDI+ inadequacices without a penalty - for example, GDI+ does not support any kind of "clamp" edge mode for gradients, only tiled/reflected modes.  This is crazy and likely a bug (https://stackoverflow.com/questions/33225410/why-does-setting-lineargradientbrush-wrapmode-to-clamp-fail-with-argumentexcepti) but I've worked around it by calculating clipping polygons on either side of the rendered gradient and manually filling them with the gradient start/stop colors on-the-fly.

I've already run a number of tests on radial gradients, however, and they pose a much bigger performance challenge.  GDI+ can render them quickly but all edge wrap options are unsupported due to the lack of a true radial gradient brush (instead we have to use a path gradient, which won't render outside the bounds of the path - argh!).  Cairo offers a wide range of helpful features here, but rendering is dog-slow at even small photo sizes.  Sigh.

And of course, conical and other "home-brew" gradient shapes are a whole other bag of worms, as we'll need to do manual per-pixel rendering for them.

What I'll likely have to do is create a new option in the rendering pipeline to inject a viewport-only preview of the current tool result, rather than our current strategy of rendering a "true" tool result in full-size in the background, and mirroring it to the viewport in real-time.

I have been hesitant to do this as I dislike when other software does it, as it's really difficult to avoid a jarring shift when the high-quality final tool result is "committed" (typically at mouse release), but I do not see another alternative at present.  Even multithreaded gradient renderers like e.g. GIMP or Paint.NET have mediocre performance and result in noticeably ugly "blocking" as results stream in.  A low-resolution preview is still better than this, IMO.

Anyway, that's why I'm committing the gradient tool in pieces rather than uploading a final "ready-to-go" version.

As the rendering pipeline changes required for additional gradient features are non-trivial to implement, I will be chipping away at them in pieces.  No ETA for a "final" version at present, sorry!