OpenRaster file format: add read support
This is a neat "first" for both PhotoDemon and VB!

The OpenRaster format is an open standard for PSD-like layered images.  It is built around ODF concepts, e.g. a zip file containing a predictable hierarchy of files.  Its current spec is here:

https://www.openraster.org/

"Good" OpenRaster support has been available in major open-source editors like GIMP, Krita, MyPaint, and others for some years (https://en.wikipedia.org/wiki/OpenRaster#Application_support).  Because my laziness keeps getting in the way of completing a PSD parser, OpenRaster is a nice stop-gap for sharing data with such programs.

This support would not be possible without cZipArchive by @wqweto.  Many thanks to him for not just providing his class under a permissive license, but also answering questions and adding features to help with this project specifically.

While the code needs to be refactored a bit for cleanliness, this initial "proof of concept" works surprisingly well!  Performance is already 4-5x faster than GIMP on comparable files and I have a few ideas for further improvement.

As admitted by the spec designers, the biggest problem with an open multi-layer image format is that programs have wildly varying support for specific features.  PD does its best to translate features from other programs where it can, but things like esoteric blend modes, layer groups, or (in MyPaint's case) non-spec inclusion of SVG vector layers, are not currently supported.  I am willing to add support for OpenRaster features as they turn up in-the-wild, provided the workload isn't insane.

OpenRaster write support is coming soon, and I expect performance to also outperform competitors (although it will be a little slower as I don't have an in-house PNG encoder yet, so I'll have to rely on 3rd-party libraries for that portion).