PSD import: support Lab color mode
8- and 16-bit channels are fully supported; like all other color modes, 32-bit channels are waiting on a proper implementation of the 32-bit "zip w/ prediction" compression decoder, which I hope to tackle soon.

Lab import is fully color-managed, so unlikely a slow, low-quality hard-coded Lab -> XYZ -> sRGB transform, PhotoDemon actually uses the embedded Lab profile in the PSD file to translate the data into an RGBA working environment.  (And if an ICC profile isn't present because the file came from a low-quality 3rd-party PSD writer, we generate a full Lab profile on-the-fly c/o lcms.)

With this, PSDs of every possible color mode are now supported.  This is a neat triumph, and much more than I thought PD would ever be capable of handling without outside help.