LittleCMS interface: expose option for converting v4 profiles to v2
When exporting PSDs, file size is not a major consideration (as the files are gonna be huge for other reasons), and v2 profiles are supported all the way back to Photoshop 5, while v4 profiles are not.  Thus embedding the slightly larger v2 profile improves compatibility in a scenario where "compatibility" is more important than any other criteria.