Overhaul internal compression interface; replace zlib with libdeflate
Lots going on in this commit - apologies for the size and no. of affected files.

PD's compression interface was originally designed against a 1:1 correlation between compression formats (e.g. DEFLATE) and the 3rd-party libraries that implement said formats (e.g. zlib).  Unfortunately, this distinction broke down when I added lz4 to the mix, as it embeds a single decompressor but two wildly different compressors (lz4 and lz4hc).  Recently this has only gotten worse as I've needed to cover both zlib and raw deflate streams, both of which are handled by zlib, but which require totally different APIs.  (Bare deflate streams require you to drop down into the zlib streaming API instead of using the single-call "block" functions.)

I've also been wanting to replace zlib with libdeflate, which offers better performance *and* compression ratios, while also providing an API that greatly simplifies working with zlib vs raw deflate data.  But this was made very messy by various parts of the program calling directly into zlib instead of even using our nice internal compression API.

All these issues meant it was time to finally overhaul PD's internal compression interface so that it operates on compression *formats* instead of *libraries*.  This required a number of changes throughout the project, typically to reroute everything through the central Compression API and to rework all functions to explicitly request a compression format as part of their compress/decompress behavior.

Long and the short of it is that PD's Compression API is now used everywhere, abstracting away the underlying libraries that implement said compression.  As an added bonus, anything that previously relied on zlib has been "silently" migrated to libdeflate.  This provides large perf improvements when working with PNG, OpenRaster, and PSD files in particular.

Bare deflate and gzip support is also available now, and I hope to make use of it in the coming days to further accelerate OpenRaster import performance.