PSD export: implement ZIP compression
There is no corresponding UI to actually activate this feature, because I need devs with modern Photoshop copies to actually test ZIP support and ensure that Adobe supports it for the large matrix of color models + color depths.  (The spec gives zero guidance here, and I've only encountered "in the wild" PSDs that use zip compression on HDR images... so I have no idea if it's allowed on SDR images, although 3rd-party readers like GIMP and Paint.NET handle the images just fine.)

Anyway, I'm committing this now in case it becomes useful to enable it in the future.