Compositor: proof-of-concept implementation of new "fast" canvas previewer
The new gradient tool is problematic, as full-canvas-size rendering is effectively impossible to implement in real-time with single-threaded VB6 constraints.  (Other FOSS editors struggle similarly, even when heavily multithreaded, so that's not a silver bullet either.)

An obvious way to mitigate this is to cheat and only render an "estimate" of the gradient during tool interaction, and when the mouse is released, invoke the necessary delay to render the "true" final gradient.  There are any number of trade-offs to an implementation like this, but it works well enough within VB6 constraints that I'm willing to try it.

This commit basically confirms that we can successfully inject a viewport-coordinate-space estimate into the rendering pipeline, and obtain a (very fast) high-quality result.  Note that it doesn't work under all situations yet - painting on vector layers, for example, requires extra work as we can't directly composite atop the vector result.  But I'm not going to spend time on that until confirming that this technique works well enough on raster layers.

Anyway, next up is implementing actual gradient rendering during the preview (instead of just a placeholder fill).